+++ srbuild/project/file_manager.py
@@ -1,6 +1,7 @@
-from srbuild.graph.graph import Graph
 from srbuild.graph.node import Node, SourceNode
+from srbuild.graph.graph import Graph
 from srbuild.logger import G_LOGGER
+
 from typing import Set, Dict, Tuple, List
 import glob
 import os