+++ srbuild/project/target.py
@@ -2,5 +2,5 @@ from srbuild.graph.node import LinkedNode
 from typing import NewType, Dict
 
 # A Dict[str, LinkedNode] that maps profile names to the LinkedNodes for a target
-class Target(dict):
+class ProjectTarget(dict):
     pass